fix bug relative to store redux not refresh

+++ src/components/ShoppingCart.jsx
@@ -1,13 +1,41 @@
 import '../Sass/components/ShoppingCartItem.scss';
 import {useDispatch} from 'react-redux';
 import * as actions from '../actions/Action';
+import { useState } from 'react';
 
 export default function ShoppingCart({ item, quantity }) {
     const dispatch = useDispatch();
-    const handleRemoveItem = (event) => {
-        dispatch(actions.removeItem(item.Id));
+    const [amount, setAmount] = useState(quantity);
+    const handleChangeQuantity = (event) => {
+       let value = event.target.value;
+       value = parseInt(value);
+        if(value < 0){
+            alert("Quantity must be  more than zero");
+            setAmount(quantity);
+            return false;
+        }
+        else if(value === 0){
+            handleRemoveItem(event);
+        }
+        else{
+            if(value > Number(item.UnitStock))
+            {
+                alert(`Quantity out of range, Unit Stock is ${item.UnitStock}`)
+                return false;
+            }
+            dispatch(actions.updateToCart(item.Id, value))
+            setAmount(value);
+        }
         event.preventDefault();
     }
+    const handleRemoveItem = (event) => {
+        let isRemove = window.confirm("Are you sure?");
+        if(isRemove){
+            dispatch(actions.removeItem(item.Id));
+        }
+        event.preventDefault();
+    }
+
     return (
         <div className="border p-2 rounded mt-2">
             <button type="button" className="close" aria-label="Close" onClick={handleRemoveItem}>
@@ -23,7 +51,7 @@ export default function ShoppingCart({ item, quantity }) {
                 </div>
                 <div className="col-2 quantity text-center">
                     <span className="text-secondary">Quantity</span><br/>
-                    <input className="form-control" value={quantity} />
+                    <input className="form-control" type="number" value={amount} onChange={handleChangeQuantity}/>
                 </div>
                 <div className="col-2 unit-price text-center">
                     <span className="text-secondary">Price</span>
@@ -31,7 +59,7 @@ export default function ShoppingCart({ item, quantity }) {
                 </div>
                 <div className="col-2 total text-center">
                     <span className="text-secondary">Total</span>
-                    <input className="form-control" value={parseInt(quantity) * parseFloat(item.Price)} />
+                    <input className="form-control" value={parseInt(amount) * parseFloat(item.Price)} />
                 </div>
             </div>
         </div>
